+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ### Added
 
 - Implement `PartialEq` and `Eq` for `Error`
+- Add an `exhaustive` modifier to `define_kind!`.
 
 ## 0.1.0 - 2026/05/22